# MindApi/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the Master Inventory Network Database (MIND) API. MIND provides automated allocation of technical attributes needed to create IP and Ethernet VPNs on the global Sigma network. MIND supports the 'Nova' services specfication which defines the collection of connectivity services supported by ENT. Go to https://thehub.thomsonreuters.com/docs/DOC-2193014 to learn more.",
    "version" : "1.0.0",
    "title" : "MIND API",
    "contact" : {
      "email" : "jonathan.beasley@thomsonreuters.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "raid.thomsonreuters.com",
  "basePath" : "/v1",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
    "/tenant" : {
      "get" : {
        "tags" : [ "tenant" ],
        "summary" : "Find all tenants",
        "description" : "Returns all tenants",
        "operationId" : "getAllTenants",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Tenant"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "tenant" ],
        "summary" : "Create a tenant",
        "description" : "This can only be done by the logged in tenant.",
        "operationId" : "createTenant",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Created tenant object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Tenant"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Tenant"
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/tenant/{tenantId}" : {
      "get" : {
        "tags" : [ "tenant" ],
        "summary" : "Find tenant by ID",
        "description" : "Returns a single tenant",
        "operationId" : "getTenantById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "description" : "ID of the tenant",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Tenant"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "tenant" ],
        "summary" : "Deletes a tenant",
        "description" : "",
        "operationId" : "deleteTenant",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "description" : "ID of the tenant",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      },
      "example" : {
        "code" : 0,
        "type" : "type",
        "message" : "message"
      }
    },
...
    "Tenant" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "tenantId" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "name" : {
          "type" : "string",
          "description" : "The name of the tenant",
          "maxLength" : 50,
          "pattern" : "^[a-zA-Z0-9-]+$"
        }
      },
      "example" : {
        "tenantId" : 0,
        "name" : "name"
      },
      "xml" : {
        "name" : "Tenant"
      }
    },
...
  },
...
}